requirements.txt
streamlit==1.41.1
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
APScheduler==3.10.4
pytest==8.3.4
pytest-mock==3.14.0
